hit1/etapa3-fault-tolerant/requirements.txt
Pillow>=10.0
numpy>=1.24
pika>=1.3.2
